+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * `initial_split`, `training`, and `testing` were added to do training/testing splits prior to resampling. 
 * `caret2rsample` and `rsample2caret` can convert `rset` objects to those used by `caret::trainControl` and vice-versa. 
+* A function called `form_pred` can be used to determine the original names of the predictors in a formula or `terms` object. 
 
 # `rsample` 0.0.1